Use normalized realm names
UnitName and GetGuildInfo return normalized realm names, but
non-normalized realm names were being used for characters on the same
realm as the player. This will always use normalized realms everywhere.

The gold tracker is an exception to avoid creating duplicate entries.
Guild duplicate entries are fine since the old guild will be removed
when the new one is created.